Extracted views from PersistentState
